fixes to make compilation by msvc 2005 possible
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@8335 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scpainterex_gdi.cpp
@@ -909,7 +909,7 @@ void ScPainterEx_GDI::drawLinearGradient( VGradientEx& gradient, const QRect& re
 
 	clipBoxWidth = rect.width();
 	clipBoxHeight = rect.height();
-	maxDim = sqrt( clipBoxWidth * clipBoxWidth + clipBoxHeight * clipBoxHeight ) * 2;
+	maxDim = sqrt( (double) (clipBoxWidth * clipBoxWidth + clipBoxHeight * clipBoxHeight) ) * 2;
 	if ( maxDim <= 0 ) 
 		return;
 
@@ -1091,7 +1091,7 @@ void ScPainterEx_GDI::drawLinearGradient_GradientFill( VGradientEx& gradient, co
 
 	clipBoxWidth = rect.width();
 	clipBoxHeight = rect.height();
-	maxDim = sqrt( clipBoxWidth * clipBoxWidth + clipBoxHeight * clipBoxHeight ) * 2;
+	maxDim = sqrt( (double) (clipBoxWidth * clipBoxWidth + clipBoxHeight * clipBoxHeight) ) * 2;
 	if ( maxDim <= 0 ) return;
 
 	x1 = m_matrix.dx() + gradient.origin().x() * m_matrix.m11();
@@ -1212,7 +1212,7 @@ void ScPainterEx_GDI::drawCircularGradient( VGradientEx& gradient, const QRect& 
 
 	clipBoxWidth = rect.width();
 	clipBoxHeight = rect.height();
-	maxDim = sqrt( clipBoxWidth * clipBoxWidth + clipBoxHeight * clipBoxHeight ) * 2;
+	maxDim = sqrt( (double) (clipBoxWidth * clipBoxWidth + clipBoxHeight * clipBoxHeight) ) * 2;
 	if ( maxDim <= 0 ) return;
 
 	FPoint pc( gradient.origin().x(), gradient.origin().y() );
